Task 02/SalesAgent/requirements.txt
python-dotenv
langchain-core
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu